requirements.txt
streamlit
streamlit-sortables
google-cloud-storage